23/08/2019 Ho creato l'Assets class, ho creato l'asset di un muro (Wall) e ho eliminato gli errori di compilazione, unico problema quando avvio il programma in modalità debug non trova l'asset (Ho fatto esattamente come su CanyonBunny)
