--- core/src/com/philosfight/game/Constants.java
@@ -1,7 +1,0 @@
-package com.philosfight.game;
-
-public class Constants {
-    // Grandezza visuale del giocatore
-    public static final float VIEWPORT_WIDTH = 5.0f;
-    public static final float VIEWPORT_HEIGHT = 5.0f;
-}
