+++ core/src/com/philosfight/game/utils/Constants.java
@@ -1,0 +1,8 @@
+package com.philosfight.game.utils;
+
+public class Constants {
+    // Grandezza visuale del giocatore
+    public static final float VIEWPORT_WIDTH = 5.0f;
+    public static final float VIEWPORT_HEIGHT = 5.0f;
+    public static final String TEXTURE_ATLAS_OBJECTS  = "images/wall";
+}
